Update how we talk about the CFRG curves
Use RFC 8032 as a reference instead of RFC 7748, as the former covers
point encodings including the sign bit, which is needed for point
addition.

This also lets us be more explicit about the procedure for generating
M and N from the infinite byte-string, and what bits to force in order
to make the byte string be interpretable as a point at all (though
checks for point validity are still needed).