Update per Hugo's review
There is not a general sense for a "strong" security proof; just say
we have a proof.

The TDH reference does not give a full proof for SPAKE2+, just an indication
for how one could prove that server compromise does not compromise plaintext
passwords; try to be more clear about that.